+++ myExpenses/src/main/java/org/totschnig/myexpenses/room/ExchangeRateDao.kt
@@ -6,7 +6,7 @@ import androidx.room.*
 interface ExchangeRateDao {
     @Query("SELECT rate from exchange_rates WHERE from_currency = :from AND to_currency = :to AND date = :date AND source = :source" +
             " UNION select 1/rate from exchange_rates where from_currency = :to and to_currency = :from AND date = :date AND source = :source limit 1")
-    suspend fun getRate(from: String, to: String, date: java.time.LocalDate, source: String): Float?
+    suspend fun getRate(from: String, to: String, date: java.time.LocalDate, source: String): Double?
 
     @Query("SELECT * from exchange_rates")
     suspend fun getAllRates(): List<ExchangeRate>
